backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
python-multipart==0.0.20
groq==0.31.0
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.8 